Base code for future MQTT Agent demo Zephyr port (#6)
* Adding interface implementation base code for mqtt_agent
This is the code that the Zephyr port will be based off of. Going to be
commiting and pushing this to main line first so the code reviews for
future MQTT Agent port for Zephyr will only need to look at changes from this base. Agent
interface is taken from
https://github.com/muneebahmed10/aws-iot-device-sdk-embedded-C/tree/mqtt-agent/demos/mqtt-agent/mqtt-agent-interface
and the subscription manager is from
https://github.com/aws/amazon-freertos/tree/main/demos/coreMQTT_Agent

* Adding base code for MQTT agent demo
Adding the code MQTT Agent demo port for Zephyr will be based off of, so
that future code reviews only need to look at changes from this code.
Some files are renamed so that git will better recognize them and
compare them with the Zephyr port.
This commit is a copy-paste of code from:
https://github.com/aws/amazon-freertos/tree/main/demos/coreMQTT_Agent
and
https://github.com/muneebahmed10/aws-iot-device-sdk-embedded-C/tree/mqtt-agent/demos/mqtt-agent