Mqtt agent demo (#7)
* mqtt_agent_demo code for Zephyr
This adds mqtt agent demo code ported for Zephyr. This does not include
the agent_interface needed for queues and the subscription manager,
which will be in another pr for the platform folder. Both prs together
should successfully build and run the mqtt agent demo.

* Changing default number of simple tasks to 1.

* Applying CR: Remove documentation meant for Windows

* Applying CR: Removing wrapper for simpleSubscribePublishTask

* Applying CR: Removing mqttAgentTaskWrapper

* Applying CR: Some changes for clarity

* Apply suggestions from code review

Fixing function not matching function declaration and voiding unused variables.

Co-authored-by: Muneeb Ahmed <54290492+muneebahmed10@users.noreply.github.com>

Co-authored-by: Muneeb Ahmed <54290492+muneebahmed10@users.noreply.github.com>